requirements.txt
cvxpy_base>=1.1
pytest

